requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
fuzzywuzzy==0.18.0
greenlet==2.0.1
idna==3.4
importlib-metadata==6.0.0
Levenshtein==0.20.9
lxml==4.9.2
packaging==22.0
Pillow==9.3.0
PyMuPDF==1.21.1
pytesseract==0.3.10
python-Levenshtein==0.20.9
rapidfuzz==2.13.7
requests==2.28.1
soupsieve==2.3.2.post1
SQLAlchemy==1.4.45
typing_extensions==4.4.0
urllib3==1.26.13
zipp==3.11.0
